Move note item into notes directory
