+++ components/notes/NoteItem.tsx
